scripts: DRY options parsing, tree printing
At this point, all tree printing in Arborist for human consumption
should be using the built in Node.toJSON printable methods.

PR-URL: https://github.com/npm/arborist/pull/216
Credit: @isaacs
Close: #216
