+++ scripts/lib/timers.js
@@ -1,0 +1,30 @@
+const ms = require('ms')
+const timers = {}
+const options = require('./options.js')
+
+process.on('time', name => {
+  if (timers[name]) {
+    throw new Error('conflicting timer! ' + name)
+  }
+  timers[name] = process.hrtime()
+})
+
+process.on('timeEnd', name => {
+  if (!timers[name]) {
+    throw new Error('timer not started! ' + name)
+  }
+  const res = process.hrtime(timers[name])
+  delete timers[name]
+
+  if (options.quiet)
+    return
+
+  console.error(name, res[0] * 1e3 + res[1] / 1e6)
+})
+
+process.on('exit', () => {
+  for (const name of Object.keys(timers)) {
+    console.error('Dangling timer: ', name)
+    process.exitCode = 1
+  }
+})
