+++ scripts/shrinkwrap.js
@@ -1,0 +1,9 @@
+const Shrinkwrap = require('../lib/shrinkwrap.js')
+const options = require('./lib/options.js')
+const { quiet } = options
+Shrinkwrap.load(options)
+  .then(s => quiet || console.log(JSON.stringify(s.data, 0, 2)))
+  .catch(er => {
+    console.error('shrinkwrap load failure', er)
+    process.exit(1)
+  })
